+++ nym-wallet/src/components/FeeWarning.stories.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react';
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { Box } from '@mui/material';
+import { BalanceWarning } from './FeeWarning';
+
+export default {
+  title: 'Wallet / Balance warning',
+  component: BalanceWarning,
+} as ComponentMeta<typeof BalanceWarning>;
+
+const Template: ComponentStory<typeof BalanceWarning> = (args) => (
+  <Box mt={2} height={800}>
+    <BalanceWarning {...args} />
+  </Box>
+);
+
+export const WithWarning = Template.bind({});
+WithWarning.args = {
+  balance: '100',
+  fee: '200',
+};
+
+export const NoWarning = Template.bind({});
+NoWarning.args = {
+  balance: '100',
+  fee: '100',
+};
